+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.8.2 (2025-08-26)
 
 - Add `Ecto.Query.API.splice/1` support https://github.com/plausible/ecto_ch/pull/240
 - Add `Ecto.Query.API.constant/1` support https://github.com/plausible/ecto_ch/pull/243